refactor(ghostvis): split ModulePartVariants resolution into GhostVisualBuilder.Variants.cs
